# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:8000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/mentorship/mentee-details/{menteeid}/{mentorshipid}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "menteeid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "mentorshipid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/mentorship/request-for-mentor": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "mentorship_industry": {
                  "example": "any"
                },
                "mentorship_topic": {
                  "example": "any"
                },
                "mentorship_goal": {
                  "example": "any"
                },
                "action_on_goal": {
                  "example": "any"
                },
                "why_a_mentor": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}